requirements.txt
django==2.2.2
django-environ~=0.4.5
pyyaml~=5.4.1
django-encrypted-model-fields~=0.5.8

# Rest Framework
djangorestframework ~= 3.12.4

# Django allauth
django-allauth==0.44.0
django-rest-auth==0.9.5

# Guthub API client
PyGithub~=1.55

# Gutlab API client
python-gitlab ~= 2.7.1

# Bitbucket API client
bitbucket-api ~= 0.5.0
atlassian-python-api ~= 3.10.0

giturlparse~=0.10.0

# Google API client
google-api-python-client~=2.11.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.4.4

# Dropbox API client
dropbox~=11.13.1
